# .futurxlab/inknowing-api-spec.yaml
openapi: 3.0.3
info:
  title: InKnowing API
  description: |
    知应(InKnowing) AI知识对话平台 MVP API规范
    基于业务逻辑守恒原理设计，确保与用户旅程、时序图、状态图完全对应
  version: 1.0.0
  contact:
    name: FuturX Lab
    email: support@inknowing.com
servers:
  - url: http://localhost:3000/api
    description: 本地开发环境
  - url: https://api.inknowing.com
    description: 生产环境
...
paths:
...
  /search:
    get:
      tags: [Books]
      summary: 智能搜索
      description: |
        自然语言搜索，AI理解用户意图
        状态转换: IDLE → SEARCHING → RESULTS_READY
        引用: #REF-SEARCH
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: 搜索关键词
      responses:
        '200':
          description: 搜索结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/BookSummary'
                  characters:
                    type: array
                    items:
                      $ref: '#/components/schemas/CharacterSummary'
                  suggestions:
                    type: array
                    items:
                      type: string
  /books:
    get:
      tags: [Books]
      summary: 获取书籍列表
      description: |
        获取所有上架书籍，支持分类和标签筛选
        状态转换: IDLE → BROWSING_BOOKS
        引用: #REF-BROWSE
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [文学, 商业, 科学, 心理, 哲学]
          description: 分类筛选
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: 标签筛选
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 书籍列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
                  total:
                    type: integer
...
components:
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        cover_url:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        ai_score:
          type: integer
        status:
          type: string
          enum: [online, offline]
    BookSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        cover_url:
          type: string
...
    CharacterSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        book_title:
          type: string
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: 用户Session认证

    adminAuth:
      type: apiKey
      in: cookie
      name: admin_session
      description: 管理员Session认证

# 业务逻辑守恒验证标记
# 所有API端点都与以下视图完全对应：
# 1. User Journey: 用户旅程中的每个动作都有对应API
# 2. Sequence Diagram: 时序图中的每个交互都有API支撑
# 3. State Diagram: 状态转换的每个触发点都是API调用
# 4. 双向可推导: 从API可以还原完整业务流程