# app/api/search/route.ts
/**
 * GET /api/search
 * 智能搜索API - 支持自然语言搜索，AI理解用户意图
 */

import { NextRequest, NextResponse } from 'next/server';
import { searchBooks } from '@/lib/services/book-service';
import { db } from '@/lib/db';
import OpenAI from 'openai';

// 通义千问客户端
const qwenClient = new OpenAI({
  apiKey: process.env.QWEN_API_KEY || '',
  baseURL: 'https://dashscope.aliyuncs.com/compatible-mode/v1',
});

/**
 * 使用AI分析搜索意图
 */
async function analyzeSearchIntent(query: string): Promise<{
  type: 'book' | 'character' | 'topic' | 'general';
  keywords: string[];
  suggestions: string[];
}> {
  try {
    const prompt = `
分析用户搜索意图：
"${query}"

返回JSON格式：
{
  "type": "book/character/topic/general",
  "keywords": ["关键词1", "关键词2"],
  "suggestions": ["相关搜索建议1", "建议2", "建议3"]
}

type说明：
- book: 搜索具体书籍
- character: 搜索角色
- topic: 搜索主题/话题
- general: 一般搜索
`;

    const completion = await qwenClient.chat.completions.create({
      model: process.env.QWEN_MODEL || 'qwen-max',
      messages: [
        { role: 'system', content: '你是一个智能搜索分析助手。' },
        { role: 'user', content: prompt }
      ],
      temperature: 0.3,
    });

    const responseText = completion.choices[0]?.message?.content || '{}';
    const jsonStr = responseText.replace(/```json\n?/g, '').replace(/```\n?/g, '');
    return JSON.parse(jsonStr);
  } catch (error) {
    console.error('AI intent analysis failed:', error);
    // 返回默认值
    return {
      type: 'general',
      keywords: [query],
      suggestions: []
    };
  }
}

/**
 * 搜索角色
 */
function searchCharacters(query: string): any[] {
  const searchPattern = `%${query}%`;

  const characters = db().prepare(`
    SELECT c.*, b.title as book_title, b.author as book_author
    FROM characters c
    INNER JOIN books b ON c.book_id = b.id
    WHERE b.status = 'published' AND (
      c.name LIKE ? OR
      c.description LIKE ?
    )
    LIMIT 10
  `).all(searchPattern, searchPattern) as any[];

  return characters.map(char => ({
    id: char.id,
    name: char.name,
    book_title: char.book_title,
    book_author: char.book_author,
    description: char.description,
  }));
}

/**
 * GET /api/search
 * 智能搜索接口
 */
export async function GET(request: NextRequest) {
  try {
    const { searchParams } = new URL(request.url);
    const query = searchParams.get('query');

    if (!query || query.trim().length === 0) {
      return NextResponse.json(
        { error: '请提供搜索关键词' },
        { status: 400 }
      );
    }

    const trimmedQuery = query.trim();

    // 使用AI分析搜索意图（可选，根据需要启用）
    let intent = { type: 'general', keywords: [trimmedQuery], suggestions: [] as string[] };

    // 如果启用AI意图分析（可根据环境变量控制）
    if (process.env.ENABLE_AI_SEARCH_INTENT === 'true') {
      intent = await analyzeSearchIntent(trimmedQuery);
    }

    // 搜索书籍
    const books = await searchBooks(trimmedQuery);

    // 搜索角色
    const characters = searchCharacters(trimmedQuery);

    // 格式化书籍结果
    const formattedBooks = books.map(book => ({
      id: book.id,
      title: book.title,
      author: book.author,
      description: book.description?.substring(0, 100) + '...',
      cover_url: book.cover_url,
      category: book.category,
      tags: book.tags,
      type: 'book',
    }));

    // 格式化角色结果
    const formattedCharacters = characters.map(char => ({
      id: char.id,
      name: char.name,
      book_title: char.book_title,
      description: char.description?.substring(0, 100) + '...',
      type: 'character',
    }));

    // 生成搜索建议
    const suggestions = intent.suggestions.length > 0
      ? intent.suggestions
      : generateDefaultSuggestions(trimmedQuery, books);

    return NextResponse.json({
      query: trimmedQuery,
      intent: intent.type,
      books: formattedBooks,
      characters: formattedCharacters,
      suggestions: suggestions,
      total: {
        books: formattedBooks.length,
        characters: formattedCharacters.length,
      },
    });
  } catch (error) {
    console.error('[Search] Error:', error);
    return NextResponse.json(
      { error: '搜索失败，请稍后重试' },
      { status: 500 }
    );
  }
}

/**
 * 生成默认搜索建议
 */
function generateDefaultSuggestions(query: string, books: any[]): string[] {
  const suggestions: string[] = [];

  // 基于找到的书籍生成建议
  if (books.length > 0) {
    const authors = [...new Set(books.map(b => b.author))];
    const categories = [...new Set(books.map(b => b.category))];

    if (authors.length > 0) {
      suggestions.push(`${authors[0]}的其他作品`);
    }
    if (categories.length > 0) {
      suggestions.push(`更多${categories[0]}类书籍`);
    }
  }

  // 默认建议
  if (suggestions.length === 0) {
    suggestions.push(
      `与"${query}"相关的书籍`,
      `搜索作者"${query}"`,
      `浏览全部书籍`
    );
  }

  return suggestions.slice(0, 3);
}